pyproject.toml
[project]
name = "essay-writer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-core>=1.0.4",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.3",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
